backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
sqlalchemy==2.0.23
pydantic==2.0.3
pydantic-settings==2.0.3
openai==1.3.5
langchain==0.0.340
chromadb==0.4.15
faster-whisper==0.9.0
gtts==2.4.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
psycopg2-binary==2.9.9
alembic==1.12.1
pytest==7.4.3
httpx==0.25.2
websockets==11.0.3
websocket-client==1.6.1
numpy==1.24.3
typing-extensions>=4.8.0
posthog==3.0.1

